Hookin' stuff up, yo
JSON support for the pkg.model.Domain type; we want it to represent
itself as having a standard string-to-string map for attributes,
even though it uses the specialized StringKVPairs type for consistent
hashing.

Hooked up the REST API to use the in-memory store for basic domain
GET and POST operations.