Inmemory store gets partition notifications
Adds an interface type to the pkg/model area, which is probably the
wrong place.

Adds a PartitionMessage container type to the pkg/model area, which is
probably a reasonable place.

Adds a JSON-oriented channel-of-channels notifier to the inmemory store,
and hooks up notifications to the source-append operation to send
partition messages through the notifier upon source appends.