Move dynamo item read into a helper type
Since we'll ultimately need to support dynamo paging, it makes sense to
move the logic for accumulating an Aggregate from dynamodb items into a
helper type that can manage the lifecycle.

This refactors the improper read implementation, pulling the query
building logic and result processing mostly out of the store proper and
into a helper type (AggregateDynamoItemReader).  The basic behavior is
unchanged for now.