Helper tools for terrible AWS exception handling
The AWS SDK doesn't appear to bother doing anything helpful with
complex error messages for batch operations against dynamodb (where
each item in the batch could produce a different result/error).

So this adds some horrors to parse that stuff out of the error
message with regexp and splits.  Delicious.